# docs/api/swagger.yaml
definitions:
...
  handlers.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/handlers.Meta'
    type: object
  handlers.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  models.Webhook:
    properties:
      created_at:
        type: string
      events:
        items:
          type: string
        type: array
      headers:
        items:
          type: string
        type: array
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  models.WebhookLog:
    properties:
      created_at:
        type: string
      duration:
        description: milliseconds
        type: integer
      error:
        type: string
      event:
        type: string
      id:
        type: integer
      payload:
        type: string
      response:
        type: integer
      success:
        type: boolean
      webhook:
        $ref: '#/definitions/models.Webhook'
      webhook_id:
        type: integer
    type: object
...
  services.CreateWebhookRequest:
    properties:
      events:
        items:
          type: string
        minItems: 1
        type: array
      headers:
        items:
          type: string
        type: array
      name:
        maxLength: 128
        type: string
      secret:
        type: string
      url:
        type: string
    required:
    - events
    - name
    - url
    type: object
...
info:
  contact: {}
paths:
...
  /webhooks:
    get:
      description: Returns all configured webhooks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Webhook'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: List webhooks
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Create a new webhook endpoint
      parameters:
      - description: Webhook config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CreateWebhookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Webhook'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Create webhook
      tags:
      - Webhooks
  /webhooks/{id}:
    delete:
      description: Delete a webhook by ID
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete webhook
      tags:
      - Webhooks
  /webhooks/{id}/logs:
    get:
      description: Returns delivery logs for a webhook
      parameters:
      - description: Webhook ID
        in: path
        name: id
        type: integer
      - default: 50
        description: Max logs
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.WebhookLog'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Webhook logs
      tags:
      - Webhooks
swagger: "2.0"